+++ cel/trunk/plugins/propclass/mover/moverfact.h
@@ -72,7 +72,7 @@ private:
     propid_moving
   };
   static Property* properties;
-  static int propertycount;
+  static size_t propertycount;
   static void UpdateProperties (iObjectRegistry* object_reg);
 
   // Normal fields.
